node tests: Move peer_data tests to new peer_data.js.
